requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# foxmulaRest/account/admin.py: 1
# foxmulaRest/account/apps.py: 1
# foxmulaRest/account/migrations/0001_initial.py: 3
# foxmulaRest/account/migrations/0002_auto_20200409_0646.py: 3
# foxmulaRest/account/models.py: 1
# foxmulaRest/account/tests.py: 1
# foxmulaRest/account/urls.py: 17,18
# foxmulaRest/account/views.py: 1
# foxmulaRest/foxmulaRest/asgi.py: 12
# foxmulaRest/foxmulaRest/urls.py: 17,18
# foxmulaRest/foxmulaRest/views.py: 1
# foxmulaRest/foxmulaRest/wsgi.py: 12
# foxmulaRest/manage.py: 10
Django == 3.0.4

# foxmulaRest/account/views.py: 4
# foxmulaRest/foxmulaRest/views.py: 4
djangorestframework == 3.11.0
